% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canning_reports.R
\name{ann_rep_chla_c}
\alias{ann_rep_chla_c}
\title{Function for annual Canning River report plots and tables for chlorophyl a.}
\usage{
ann_rep_chla_c(outpath, data, surface, chloro)
}
\arguments{
\item{outpath}{filepath to desired export location.}

\item{data}{the data object created from running `canning_WIN_report_data`.}

\item{surface}{colour for surface plots. Can be named colour or hex format.}

\item{chloro}{colour for integrated plots. Can be named colour or hex format.}
}
\value{
a  panel plot for chlorophyl a and a csv of metrics for
    inclusion to a table.
}
\description{
\code{ann_rep_chla_c} takes an export filepath, data, surface and integrated
    plot colours and produces panel plots and tables for chlorophyl a.
}
\details{
Outputs will be exported to two folders  at the outpath
    location. `c_panels/` for plots and `c_tables/` for data tables. These
    will need to be created if they don't exist.
}
\examples{
\dontrun{
ann_rep_chla_c(outpath, data, surface = "blue", chloro = "darkgreen")}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
