% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swan_reports.R
\name{ann_rep_grp2_s}
\alias{ann_rep_grp2_s}
\title{Function for annual Swan River report plots and tables for group 2
metrics}
\usage{
ann_rep_grp2_s(outpath, data, surface)
}
\arguments{
\item{outpath}{filepath to desired export location.}

\item{data}{the data object created from running `swan_WIN_report_data`.}

\item{surface}{colour for surface plots. Can be named colour or hex format.}
}
\value{
a separate panel plot for each metric and a csv of metrics for
    inclusion to a table.
}
\description{
\code{ann_rep_grp2_s} takes an export filepath, data, and surface plot
    colours and produces panel plots and tables for group 2 metrics.
}
\details{
Group 2 metrics are a grouping based on a common collection
    method, sample type and display parameters and in this case includes,
    alkalinity, dissolved organic carbon and TSS. Outputs will be exported
    to two folders at the outpath location. `s_panels/` for plots
    and `s_tables/` for data tables. These will need to be created if they
    don't exist.
}
\examples{
\dontrun{
ann_rep_grp2_s(outpath, data, surface = "blue")}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/rivRmon/index.html}
{the rivRmon website}
}
