% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canning_reports.R
\name{ann_rep_grp3_c}
\alias{ann_rep_grp3_c}
\title{Function for annual Canning River report plots and tables for group 3
metrics}
\usage{
ann_rep_grp3_c(outpath, data, surface, bottom)
}
\arguments{
\item{outpath}{filepath to desired export location.}

\item{data}{the data object created from running `canning_WIN_report_data`.}

\item{surface}{colour for surface plots. Can be named colour or hex format.}

\item{bottom}{colour for bottom plots. Can be named colour or hex format.}
}
\value{
a separate panel plot for each nutrient and a csv of metrics for
    inclusion to a table.
}
\description{
\code{ann_rep_grp3_c} takes an export filepath, data, surface and bottom
    plot colours and produces panel plots and tables for group 3 metrics.
}
\details{
Group 3 metrics are a grouping based on a common collection
    method, sample type and display parameters and in this case includes,
    dissolved oxygen, salinity, temperature, and pH.
    Outputs will be exported to two folders at the outpath location.
    `c_panels/` for plots and `c_tables/` for data tables. These will need to
    be created if they don't exist.
}
\examples{
\dontrun{
ann_rep_grp3_c(outpath, data, surface = "blue", bottom = "red")}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://Bartesto.github.io/rivRmon/index.html}
{the rivRmon website}
}
