% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swan_reports.R
\name{ann_rep_secchi_s}
\alias{ann_rep_secchi_s}
\title{Function for annual Swan River report plots and tables for secchi depth.}
\usage{
ann_rep_secchi_s(outpath, data, surface)
}
\arguments{
\item{outpath}{filepath to desired export location.}

\item{data}{the data object created from running `swan_WIN_report_data`.}

\item{surface}{colour for surface plots. Can be named colour or hex format.}
}
\value{
a  panel plot for secchi depth and a csv of metrics for
    inclusion to a table.
}
\description{
\code{ann_rep_secchi_s} takes an export filepath, data, surface plot colours
    and produces panel plots and tables for secchi depth.
}
\details{
Outputs will be exported to two folders at the outpath location.
    `s_panels/` for plots and `s_tables/` for data tables. These will need to
    be created if they don't exist.
}
\examples{
\dontrun{
ann_rep_secchi_s(outpath, data, surface = "blue")}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://Bartesto.github.io/rivRmon/index.html}
{the rivRmon website}
}
