% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canning_reports.R
\name{ann_rep_si_c}
\alias{ann_rep_si_c}
\title{Function for annual Canning River report plots and tables for silica.}
\usage{
ann_rep_si_c(outpath, data, surface)
}
\arguments{
\item{outpath}{filepath to desired export location.}

\item{data}{the data object created from running `canning_WIN_report_data`.}

\item{surface}{colour for surface plots. Can be named colour or hex format.}
}
\value{
a  panel plot for silica and a csv of metrics for
    inclusion to a table.
}
\description{
\code{ann_rep_si_c} takes an export filepath, data, and surface
    plot colour and produces panel plots and tables for silica.
}
\details{
Outputs will be exported to two folders at the outpath
    location. `c_panels/` for plots and `c_tables/` for data tables. These
    will need to be created if they don't exist.
}
\examples{
\dontrun{
ann_rep_si_c(outpath, data, surface = "blue")}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
