% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKAN_canning_surfer.R
\name{CKAN_canning_surfR}
\alias{CKAN_canning_surfR}
\title{Reads in external data, combines with internal datasets and produces Canning
River surfer plots}
\usage{
CKAN_canning_surfR(path, obac, onic)
}
\arguments{
\item{path}{Character string filepath to location of sonde data xlsx
workbooks.}

\item{obac}{Character string ("green", "blue" or "red") indicating
oxygenation plant status at site BAC.}

\item{onic}{Character string ("green", "blue" or "red") indicating
oxygenation plant status at site NIC.}
}
\value{
A pdf format four panel surfer plot of the Canning River. Saved to
    a directory called `plots/`.
}
\description{
\code{canning_surfR} takes a file path to Canning River sonde output and creates
    a four panel (single column) surfer plot of salinity, dissolved oxygen,
    chlorophyll a, and temperature in pdf format. The function creates a
    directory called `plots/` in the file path to store the pdf's. Code expects
    only 2 excel workbooks for one monitoring run.  Note sonde data from EXO
    models reads depth from the VPos metric. All other sonde models use a
    variant of a depth metric.
}
\details{
Surfer plots display a
    cross-section of the river where the metrics of interest have been
    interpolated between sonde locations. Thin plate spline has
    been used for the interpolation.

    A river "bottom" is displayed which puts the interpolation in context.
    The river bottom has been derived from a combination of historical
    maximum depths at sampling locations as well as station points extracted
    from a "best" navigation line over the latest corporate bathymetry. See
    project documentation for further details.
}
\examples{
\dontrun{
canning_surfR(path = "Z:/DEC/MonitoringProgram/Data", obac = "green", onic = "red")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
