% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKAN_swan_surfer.R
\name{CKAN_swan_surfR}
\alias{CKAN_swan_surfR}
\title{CKAN version. Reads in external data, combines with internal datasets and produces Swan
River surfer plots}
\usage{
CKAN_swan_surfR(path, ovit, ocav)
}
\arguments{
\item{path}{Character string filepath to Project location. Sonde data xlsx
workbooks will be there in a directory within the Project called `data/`.}

\item{ovit}{Character string ("green", "blue" or "red") indicating
oxygenation plant status at site VIT.}

\item{ocav}{Character string ("green", "blue" or "red") indicating
oxygenation plant status at site CAV.}
}
\value{
Two pdf format four panel surfer plots of the Swan River. One shows
   the full extent of the monitoring run (river mouth to the just beyond the
   site POL - Upper Swan Power Lines). The second shows from the Narrows
   Bridge to site POL. Saved to Project directory `plots/`.
}
\description{
\code{CKAN_swan_surfR} takes a file path to a Project that has a folder
    called "data" that contains Swan River sonde outputs and creates
    a four panel (single column) surfer plot of salinity, dissolved oxygen,
    chlorophyll a, and temperature in pdf format. Code expects
    only 2 excel workbooks for one monitoring run. Note sonde data from EXO
    models reads depth from the VPos metric. All other sonde models use a
    variant of a depth metric.
}
\details{
Surfer plots display a
    cross-section of the river where the metrics of interest have been
    interpolated between sonde locations. Thin plate spline has
    been used for the interpolation.

A river "bottom" is displayed which puts the interpolation in context.
    The river bottom has been derived from a combination of historical
    maximum depths at sampling locations as well as station points extracted
    from a "best" navigation line over the latest corporate bathymetry. See
    project documentation for further details.
}
\examples{
\dontrun{
swan_surfR(path = "Z:/DEC/MonitoringProgram", ovit = "green", ocav = "red")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
