% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.r
\name{data_finder}
\alias{data_finder}
\title{Finds filepath to specific river data}
\usage{
data_finder(path, river)
}
\arguments{
\item{path}{Character string filepath to location of sonde data xlsx
workbooks.}

\item{river}{Character. Either "c" for Canning River or "s" for Swan
River.}
}
\value{
If data meeting the required conditions exist, filepaths of data
    location. There should be two filepaths per date for upper and lower
    data runs.
}
\description{
\code{data_finder} takes a filepath and a river designator and returns the
    complete filepaths to specified river data.
}
\details{
The function presumes data to be in xlsx format and the river designator,
within the name of the file, to be preceded by 8 digits (date).
}
\examples{
\dontrun{
data_finder(path = "Z:/DEC/MonitoringProgram/Data", river = "c")}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
