% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_report_functions.R
\name{dynamic_ylim}
\alias{dynamic_ylim}
\title{Helper function to calculate sensible y axis limits from the data}
\usage{
dynamic_ylim(x)
}
\arguments{
\item{x}{the data.}
}
\value{
a numeric vector of length 2, first value zero, second the calculated
    y limit. Will be passed to ylim argument in a ggplot call.
}
\description{
\code{dynamic_ylim} takes the data for a variable and calculates the maximum
    value of the background and present period plotting variables, finds the
    maximum of those values and adds 10%. This bcomes the upper y limit.
    Lower y limit is zero. Is highly very specific to data structure created
    when processing the annual report data.
}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
