% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_report_functions.R
\name{get_legend}
\alias{get_legend}
\title{Helper function for extracting legend from a plot}
\usage{
get_legend(myggplot)
}
\arguments{
\item{myggplot}{a ggplot object that contains a legend}
}
\value{
a legend grob
}
\description{
\code{get_legend} takes a ggplot object and returns the legend grob
}
\examples{
\dontrun{
get_legend(myggplot)}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
