% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hab_functions.R
\name{hab_tablR}
\alias{hab_tablR}
\title{Creates conditionally formatted table as a pptx file from summary HAB data}
\usage{
hab_tablR(hab_tables, focus_date, mngt_triggers)
}
\arguments{
\item{hab_tables}{a character file path to the location of the `HAB_tables/`
directory.}

\item{mngt_triggers}{a character filepath and name of the 
`mngt_response_triggers_YYYY.csv` data file that indicates group names, 
species names and trigger thresholds for management action. See package
example dataset `mngt_response_triggers` for example of format required.}

\item{date}{a character representation of the date chosen for the "present"
sampling week in the format "yyyymmdd".}
}
\description{
\code{hab_tablR} takes a file path to summary HAB data and a date and produces
    "present" and "prior sample" table of phytoplankton counts per site.
}
\details{
This function can only be run on data created by running the
    (\code{\link{hab_groupR}}) function. It will read in all the summary
    data then based on the provided date, produce a conditionally formatted 
    table for that date. Values are coloured according to "investigate" or 
    "notification" count level for river sites for specific species. The prior
    sample's data is reported so comparisons can be made.

    Data is exported as pptx file to the `HAB_tables/` directory. It can be 
    dragged and dropped into a PowerPoint template with some minor resizing.
}
\examples{
\dontrun{
hab_tablR(summary = "C:/path/to/HAB_tables", date = "20190910",
mngt_triggers = "./mngt_response_triggers_2020.csv")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://Bartesto.github.io/rivRmon/index.html}
{the rivRmon website}
}
