% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-triggers.R
\docType{data}
\name{mngt_response_triggers}
\alias{mngt_response_triggers}
\title{Management Response Triggers}
\format{
a data.frame with 5 variables:
\describe{
\item{table_ord}{numeric index of the desired order of reportable groups}
\item{table_name}{character representing the reportable group name as it is intended for the output}
\item{spp_name}{character representing the exact species name/s that compromise the reportable group}
\item{surveillance}{numeric values below this are categorised as surveillance}
\item{notification}{numeric values equal to or above are categorised as notification}
}
}
\usage{
mngt_response_triggers
}
\description{
This data set is an example of the format required for creating a table for 
ingestion to the (\code{\link{hab_groupR}}) and (\code{\link{hab_tablR}}) 
functions. Values in this dataset determine which species are summarised and 
what values constitute differing levels of management action.
}
\details{
Although this is an example of the current triggers (as at time of writing) it 
is intended that this be used as a template to guide the creation of trigger 
data that would be saved out as a csv for use with the HAB functions. Note 
that values falling between surveillance and notification are classified as 
investigate.
}
\examples{
mngt_response_triggers
}
\keyword{datasets}
