% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.r
\name{phyto_finder}
\alias{phyto_finder}
\title{Creates file paths to raw phytoplankton xlsx spreadsheets}
\usage{
phyto_finder(pathin)
}
\arguments{
\item{pathin}{a character filepath to the location of the raw xlsx
spreadsheets}
}
\description{
\code{phyto_finder} takes a file path to raw phytoplankton xlsx spreadsheets
    and returns a full filepath for use in other functions.
}
\details{
This is a helper function to set up correct file paths to the
    appropriate data xlsx spreadsheets.
}
\examples{
\dontrun{
phyto_finder(pathin = "C:/path/to/raw_data")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/rivRmon/index.html}
{the rivRmon website}
}
