% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyto_functions.R
\name{phyto_plotR}
\alias{phyto_plotR}
\title{Creates barcharts from phytoplankton summary data}
\usage{
phyto_plotR(summary, date, SHELL = FALSE)
}
\arguments{
\item{summary}{a character file path to the location of the `summaries/`
directory.}

\item{date}{a character representation of the date chosen for the "present"
barchart in the format "yyyymmdd".}
}
\description{
\code{phyto_plotR} takes a file path to summary data and a date and produces
    "present" and "prior sample" barcharts of phytoplankton counts per site.
}
\details{
This function can only be run on data created by running the
    (\code{\link{phyto_groupR}}) function. It will read in all the summary
    data then based on the provided date, produce a barchart for that date
    and the prior sample's summary in the one output.

    As data is gathered for each estuary on consecutive days, the choice of
    date will determine which river summaries will be plotted. That choice
    will be reflected in the sites plotted in the barcharts. These barcharts
    will be exported (as pdf) to the `summaries/` directory.

    Lastly this function will export a "running_datasheet" that contains all
    summary data for the estuary prior to and including the date requested.
}
\examples{
\dontrun{
phyto_plotR(summary = "C:/path/to/summary_data", date = "20190910")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://Bartesto.github.io/rivRmon/index.html}
{the rivRmon website}
}
