% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.r
\name{sonde_reader}
\alias{sonde_reader}
\title{Reads in sonde data from various model formats and creates tidy data}
\usage{
sonde_reader(path)
}
\arguments{
\item{path}{Character string filepath to location of sonde data xlsx
workbook (one only).}
}
\value{
A data frame containing:
    \itemize{
      \item site
      \item temperature
      \item dissolved oxygen mg/L
      \item salinity ppt
      \item depth
      \item chlorophyll a
      }
}
\description{
\code{sonde_reader} takes a filepath to one xlsx workbook as identified
    by \code{\link{data_finder}} and determines:
    \itemize{
      \item determines whether data has two header rows
      \item cleans data
      \item returns a tidy data set
      }
}
\examples{
\dontrun{
sonde_reader(path = "Z:/DEC/MonitoringProgram/Data/20191201cpoEXO1.xlsx")}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
