% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.r
\name{sonde_reader_metric}
\alias{sonde_reader_metric}
\title{Reads in sonde data and formats specifically for use for plotting single metrics}
\usage{
sonde_reader_metric(path)
}
\arguments{
\item{path}{Character string filepath to location of manually combined sonde 
data xlsx workbook (one only). Looks for shhet tab beginning with the letter 
'e'.}
}
\value{
A data frame where the 'VPos' variable has been renamed to "Dep m" 
   to allow for consistency with legacy code.
}
\description{
\code{sonde_reader_metric} takes a filepath to one xlsx workbook as identified
   by \code{\link{data_finder}} and formats it for single metric plotting.
}
\examples{
\dontrun{
sonde_reader_metric(path = "Z:/DEC/MonitoringProgram/Data/20191201cpoEXO1.xlsx")}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/rivRmon/index.html}
{the rivRmon website}
}
