% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.r
\name{surfer_cols}
\alias{surfer_cols}
\title{Creates correct colour palette for measured metric}
\usage{
surfer_cols(metric)
}
\arguments{
\item{metric}{Character string, "sal" for Salinity, "do" for Dissolved
Oxygen mg/L, "chl" for Chlorophyll a, or "temp" for Temperature.}
}
\value{
A named character vector where names relate to data break values and
   characters are colour hex codes.
}
\description{
\code{surfer_cols} takes metric designator and returns a named charcater
    of colours.
}
\examples{
colours <- surfer_cols(metric = "chl")

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
