% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_report_functions.R
\name{swan_WIN_report_data}
\alias{swan_WIN_report_data}
\title{Function to read and format water quality data obtained from the WIR water
portal for use in annual reporting for the Swan River}
\usage{
swan_WIN_report_data(inpath, reportingYear = 2019, outpath)
}
\arguments{
\item{inpath}{character filepath to the downloaded data xlsx files.}

\item{reportingYear}{4 digit numerical representing current reporting year.}

\item{outpath}{character filepath to desired export location.}
}
\value{
a csv export named SG-E-SWANEST_annual_report_data_for_YYYY.
}
\description{
\code{swan_WIN_report_data} reads in all xlsx downloaded files from WIR
    and based on a reporting year, formats the data for use in generating the
    annual report.
}
\details{
Original source data must be downloaded from
    \url{http://wir.water.wa.gov.au/Pages/Water-Information-Reporting.aspx}
    using the keyword "SG-E-SWANEST" and requesting "Water Quality (discrete)
    for Site(s) cross-tabulated". These datasets are then read in and
    reformatted for ease of use in the annual report.
}
\examples{
\dontrun{
swan_WIN_report_data(inpath, reportYear = 2019, outpath)}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
